==> Code/2_beta_diversity.Rmd <==
---
title: "2_beta_diversity"
author: "Pedro Beschoren da Costa"
date: "August 25, 2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

# load libraries and data
```{r}
source("./Code/Functions/load_libraries.R")
load("./R output/env.1_Loading_and_pre_processing.RData")
```




# 2.0 - basic beta diversity analysis
On this script we will evaluate beta diversity - the differences in the microbial community compositions across samples. we will make ordination plots and multivariate tests


## 2.1 - Beta Diversity plots
Beta diversity plots are the beating heart or microbiome analysis. here you will be able to visually tell if communities differ according to treatment or not. It can be a very long topic but here I only use one option.


```{r}

# Let's make a Non-Metric Multidimensional Scaling (NMDS) of all our samples based on CSS normalization
nmds_rootAndSoil <- phyloseq::ordinate(physeq_filtered_CSS,
  method = "NMDS", # this method has few assumptions and readly accepts different data structures
  distance = "bray", # bray-curtis distance is suitable for sparse data - such as the zero-inflated microbiome data we have
  try = 200, # number of iterations
  autotransform = TRUE
) # automatically transforms your data, if needed. reduces weight of outliers
# weakties = FALSE prevests tress from colapsing into zero

plot_ordination(
  physeq = physeq_filtered_CSS,
  ordination = nmds_rootAndSoil,
  color = "MeJA_treatment",
  shape = "Plant_species"
) +
  theme_classic() +
  theme(plot.title = element_text(size = 10, face = "bold", hjust = 0.5)) +
  geom_point(aes(size = Sample_type), alpha = 1) +
  theme(legend.position = "right")

metadata <- as(sample_data(physeq_filtered_CSS), "data.frame")
```


### 2.1.1 - custom NMDS function for automation of split datasets

```{r}

# define custom fucntion to plot a list of NMDS results
NMDS_listing <- function(physeq_list) { # first the name of the new function you are saving, then you say you are going to define a function, and then you set it's arguments - here, a single list of phyloseq objects. the {} indicate where the contents of the custom function start and end.

  # here we perform the NMDS on all elements of the list using phyloseq::ordinate()
  NMDS_list <- lapply(physeq_list, function(x) {
    ordinate(
      physeq = x, # your phyloseq object, each element x in the list
      method = "NMDS", # ordination method to use
      distance = "bray", # distance metric to use
      try = 200, # number of permutations
      autotransform = TRUE
    )
  }) # automatic data transformation

  # mapply will perform one function on two lists. define the lists in the end of the function
  untitled_plot_list <- mapply(function(x, y) { # mapply will run one function on 2 lists
    plot_ordination(
      physeq = x, # your phyloseq object, each element x in the list
      ordination = y, # your phyloseq-generated ordination object, each element y in the list made above
      color = "MeJA_treatment"
    ) + # dot colors in the plot
      theme_classic() + # ggplot option to remove background
      labs(subtitle = paste("Stress:", round(y$stress, digits = 4))) + # this adds the NMDS stress to the plot as subtitle
      theme(plot.title = element_text(size = 10, face = "bold")) + # options for the title
      theme(legend.position = "right")
  }, # position for the legend
  x = physeq_list, # note that you only define x here, after you define your plot function
  y = NMDS_list, # note that you only definee y here, after you define your plot function
  SIMPLIFY = FALSE
  ) # if you simply, all resusts are saved in a vector

  # The plots above miss a title, so you don't really know who is root and who is soil. this code create a list of names to be used as titles in your plot
  tiles_list <- names(untitled_plot_list)

  Plot_list <- mapply(function(z, w) { # use mapply again to put those listed names in the list of plots
    z + ggtitle(w)
  },
  z = untitled_plot_list,
  w = tiles_list,
  SIMPLIFY = FALSE
  )

  return(list(NMDS_list, Plot_list)) # with this you specify the output of your custom function. Note we are saving both the NMDS calculations and the plot as a list, so we don't ahve to re-calculate the NMDS later
}

# let's run our custom function
set.seed(5235)
NMDS_4_slice <- NMDS_listing(ps_list_CSS)
```


## 2.2 - Beta Dispersion

Beta dispersion indicates variance in the multidimensional space. It essentially measures the spread of your data points in the ordination, according your different treatments. As you might expect, statistical tests will have a hard time comparing a treatment with high beta dispersion (wide, sparse datapoint cloud) against a treatment with low beta dispersion (clustered, compact datapoint cloud), just like it is difficult to compare large error bars to small error bars.  You can test for homogeniety of this variance, similarly to a levene test of homogeniety before an anova. 

What if you don't have homogeniety of variances? then your permanova results are not as reliale. This can be difficult to solve. there are a few transformations you could try (log or sqrt of your counts), outliers you could remove (specially if you observed that a particular outlier was specially unhealthy) or just split your permanovas accordingly (if soil samples have higher beta dispersion than root samples, test root and soil in separate permanovas). sometimes however these alternatives do no solve the problem - proceed with your tests as normal but be cautious with your conclusions.

with this chunk we evaluate Beta Dipersion Metrics & plots 
```{r}

# this will test beta dispersions over a list by considering a factor
#### -------- soft-coded, 3 lists of phyloseq objects ---###
beta_disp_plotAndTest <- function(phyloseq_list, group) {
  # phyloseq_list = a list of phyloseq objects
  # group = the variale you want to test the beta dispersion of, in quotes
  beta_disper_list <- lapply(phyloseq_list, function(x) {
    betadisper(phyloseq::distance(t(otu_table(x)), method = "bray"), sample_data(x)[[group]])
  }) # selects only column "group""

  # gets the names of the list
  tiles_list <- names(beta_disper_list)

  # runs anova on beta dispersions
  get_p <- lapply(beta_disper_list, function(x) {
    anova(x, permutations = 999)
  })


  p_dispersion <- map(get_p, 5) # gets the p value of the dispersion test
  p_dispersion <- p_dispersion[!is.na(p_dispersion)] # removes the generated NA


  # runs anova on beta dispersions
  bet_disp_PCOa_plot <- mapply(function(x, y, z) {
    plot(x,
      main = y,
      sub = z,
      xlab = "p value for homogeniety test:",
      ylab = "PCoA"
    )
  },
  x = beta_disper_list,
  y = tiles_list,
  z = p_dispersion,
  SIMPLIFY = FALSE
  )

  # runs anova on beta dispersions
  bet_disp_boxplot <- mapply(function(x, y) {
    boxplot(x, main = y)
  },
  x = beta_disper_list,
  y = tiles_list,
  SIMPLIFY = FALSE
  )


  return(list(bet_disp_PCOa_plot, bet_disp_boxplot))
}

# now that we have a custom function, we can run it across all lists and variables
set.seed(5235)
dip_result3 <- beta_disp_plotAndTest(ps_list_CSS, "MeJA_treatment")
```



## 2.3 - Run PERMANOVA 

A permutation anova will tell if the differences in the microbial community structure are significant or not. they will essentially help you separate the data clouds of your ordination with confidence levels


you will need to run, test and check several different models and data slices to have final insight into the dataset you are evaluating. get used with testing multiple models!

```{r}
# Running the permanova with vegan::adonis2() on a single phyloseq object is very simple
metadata <- as(sample_data(physeq_filtered_CSS), "data.frame")
# let's first define permitation design with permute::how()
perm2 <- how(
  within = Within(type = "free"), # permute observations freely within the blocks. got to ask Fred how to handle grid arrangements!
  blocks = metadata$Block, # define blocks
  nperm = 999
) # n of permutations
set.seed(5235)
# now let's calculate the permanova
three_way_permanova <- adonis2(phyloseq::distance(t(otu_table(physeq_filtered_CSS)), method = "bray") # this is your distance matrix of OTU abundances
~ MeJA_treatment * Plant_species * Sample_type + Block, # this is your model
permutations = perm2, # we jsut defined the permutation settings above, you place them here
data = metadata
) # metadata for the community matrix



#### -------- soft-coded, 3 lists of phyloseq objects with 1 formula ---###

# let's check now if sample type differ per species. note that this function uses Blocks as strata
permanova_with_blocks <- function(phyloseq_list, rhs_model) {
  # phyloseq_list = list of phyloseq objects
  # RHS_model = right hand side model, such as MeJA_treatment*Sample_type + Block
  lapply(phyloseq_list, function(x) {
    lhs_data <- phyloseq::distance(t(otu_table(x)), method = "bray")
    rhs_model_in <- paste(rhs_model)
    form <- as.formula(paste("lhs_data~", paste(rhs_model_in))) # getting the formulat properly evaluated as custom string is tricky
    output <- adonis2(form,
      data = as(sample_data(x), "data.frame"), # changing with as.data.frame is insufficient
      permutations = how(within = Within(type = "free"), blocks = as(sample_data(x), "data.frame")$Block, nperm = 9999)
    ) # how defines the permutations, it is important to adjust it to the experimental design such as a time series
    return(output)
  })
}

# run custom funcion
set.seed(5235)
permanova_with_blocks(ps_list_CSS, "MeJA_treatment + Block")


```
## 2.4 pairwise PERMANOVA

Now that we know the fixed factor effects and interactions, let's make pairwise comparisons.for this we will use EcolUtils::adonis.pair - with lists and p adjust, no blocks

```{r}

# pairwise coparisons over a list fo 4 ps obects
set.seed(5235)
lapply(ps_list_CSS, function(x) {
  adonis.pair(
    dist.mat = phyloseq::distance(otu_table(x), method = "bray"),
    Factor = as.factor(as(phyloseq::sample_data(x), "data.frame")$MeJA_treatment)
  )
})


# pairwise coparisons over afull community
adonis.pair(
dist.mat = phyloseq::distance(otu_table(physeq_filtered_CSS), method = "bray"),
Factor = as.factor(as(phyloseq::sample_data(physeq_filtered_CSS), "data.frame")$MeJA_treatment)
)



```


# 2.5 - Save session
```{r}

save.image("./R output/analysis_session.RData")
```


Beta Diversity analysed! proceed to scrip 3_Alpha_Diversity!
