==> Code/3_alpha diversity.Rmd <==
---
title: "3_alpha diversity"
author: "Pedro Beschoren da Costa"
date: "August 27, 2021"
output: html_document
editor_options: 
  chunk_output_type: console
---


# load libraries and data
```{r}
source("./Code/Functions/load_libraries.R")
load("./R output/analysis_session.RData")
```




# 3.0 Basic alpha diversity analysis for pilot data
Aplha diversity will show you the quantifiable diversity within a sample - this way you compare which treatment is more or less diverse or species-rich than the other. Like in all the other steps we are evaluating here, there are countless variations and methods you could use. this is just a convenient template for you to get started.

Alpha diversity requires counts from samples with equal samples sizes. this means we cannot use the CSS transformation from the metagenomeseq package. also, heavy filtering of your data may skew some of the diversity metrics (due to the lack of rare species), so we should not over-filter the data before calculating alpha diversity.

## 3.1 simple Alpha diversity plots
```{r}

# as noted above, we will use rarefied data, filtered at at least 8 occurences in the dataset
# This is a very simple plot, but will serve for a visualization
plot_richness(physeq_filtered_rarefied, x = "MeJA_treatment", measures = c("Observed"), color = "Plant_species") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90)) +
  labs(color = "Plant species") +
  ylab("Observed number of ASVs") +
  facet_wrap(~Sample_type)
# as we could already expect, soils are more diverse than roots. for Brassica However the difference, if significant, is very small
```
## 3.1 Alpha diversity testing

Here we check shanon diversity indexes
```{r}
# Calculate richness for root and soil
total_diversity <- estimate_richness(physeq_filtered_rarefied) %>%
  rownames_to_column(var = "Sample")
total_diversity$Sample <- sub("X", "", total_diversity$Sample) # why the hell that X popped up? this fix it
total_diversity <- column_to_rownames(total_diversity, var = "Sample")


#### add diversity metrics to mapping file of phyloseq objects
# we do this so we can perform anovas, acess metadat, make nicer plots, etc
merg_to_ps <- sample_data(total_diversity) # makes the diversity calculations  sample_data for phyloseq oject...
sample_data(physeq_filtered_rarefied) <- merge_phyloseq(physeq_filtered_rarefied, merg_to_ps) # merge the new phyloseq object with the old phyloseq object
total_diversity <- as(sample_data(physeq_filtered_rarefied), "data.frame") # forces sample data of updated phyloseq object into a dataframe


# check homogeniety of variances
leveneTest((Shannon) ~ Plant_species * MeJA_treatment * Sample_type, data = total_diversity)

# two-way anova, shanon diversity index
tx <- with(total_diversity, interaction(MeJA_treatment, Plant_species, Sample_type)) # needed for tukey to test interaction
aovTukey <- aov(Shannon ~ tx, data = total_diversity) # needed for tukey to test interaction
Anova(lm((Shannon) ~ Block + Plant_species * MeJA_treatment * Sample_type, data = total_diversity, contrasts = list(MeJA_treatment = contr.sum, Plant_species = contr.sum)), type = "2")
HSD.test(aovTukey, "tx", group = TRUE, console = TRUE) # post-hoc

# Lets see this on a simple boxplot for Shannon diversity index
plot_alpha_BySample <- ggplot(total_diversity, aes(x = MeJA_treatment, y = Shannon, fill = Plant_species)) +
  geom_boxplot() +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(y = "Shannon diversity index") +
  theme(
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank()
  ) +
  facet_wrap(~Sample_type)

```

## 3.2 Alpha diversity testing on individual data parittions


Now that we see clear diversity effects, let's check in details to detect key pairwise differences wiht lists and loops


```{r}


# this will add the diversity metrics to a phyloseq object
add_diversity_to_physeq_object <- function(phyloseq_object) {

  # Calculate richness for root and soil of a sigle phyloseq object
  total_diversity <- estimate_richness(phyloseq_object) %>%
    rownames_to_column(var = "Sample")

  total_diversity$Sample <- sub("X", "", total_diversity$Sample) # why the hell that X popped up? this fix it
  total_diversity <- column_to_rownames(total_diversity, var = "Sample")


  #### add diversity metrics to mapping file of phyloseq objects
  # we do this so we can perform anovas, acess metadat, make nicer plots, etc
  merg_to_ps <- sample_data(total_diversity) # makes the diversity calculations  sample_data for phyloseq oject...
  sample_data(phyloseq_object) <- merge_phyloseq(phyloseq_object, merg_to_ps) # merge the new phyloseq object with the old phyloseq object
  diversity <- as(sample_data(phyloseq_object), "data.frame") # forces sample data of updated phyloseq object into a dataframe

  return(diversity)
}


# runs new custom function on a list of phyloseq objects, making a list of dataframes with alpha diversity indexes
diversity_list <- lapply(ps_list_rarefied, add_diversity_to_physeq_object)

# check levenes homogeniety over the list
lapply(diversity_list, function(x) {
  leveneTest((Shannon) ~ MeJA_treatment, data = x)
})

# since our initial object has been sliced into multiple parts, the anova is much simpler
# build model
aov_list <- lapply(diversity_list, function(x) {
  aov(lm((Shannon) ~ Block + MeJA_treatment, data = x))
})

# run anova
lapply(aov_list, anova)

# check tukey
lapply(aov_list, function(x) {
  TukeyHSD(x, "MeJA_treatment")
})


```




# 3.3 - overwrite saved session
```{r}

save.image("./R output/analysis_session.RData")
```




Alpha Diversity analyised! proceed to scrip 4_neutral_models!
