==> Code/7_random_forest.Rmd <==
---
title: "7_random_forest"
author: "Pedro"
date: "10/26/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

# load libraries and data
```{r}
source("./Code/Functions/load_libraries.R")
load("./R output/analysis_session.RData")
```




# 7.0 - Random forest
I've used this tutorial https://www.andreaperlato.com/mlpost/feature-selection-using-boruta-algorithm/
you can also conside rthis: https://www.analyticsvidhya.com/blog/2016/03/select-important-variables-boruta-package/
https://www.analyticsvidhya.com/blog/2021/05/boruta-with-r-is-a-great-way-to-address-the-curse-of-dimensionality/



## 7.1 - using borutra over a list of phyloseq objects
```{r boruta}


# prepare phyloseq object to be an input in Boruta
physeq_to_borutaInput <- function(list_physeq_object, variable_to_be_classified) {
  # key input: transposed OTU table with Meja_treatment as a added variable
  # the output is a list of df ready to be used as input to boruta
  # transpose phtseq otu table
  otu_cells_wide_list <- lapply(list_physeq_object, function(x) { # transpose the feature table...
    as.data.frame(t(otu_table(x))) %>%
      rownames_to_column(var = "sample")
  })

  # extract sample data
  metadata_list <- lapply(list_physeq_object, function(x) {
    as(sample_data(x), "data.frame") %>%
      rownames_to_column(var = "sample")
  })

  # add the variable classification you want to predict with the random forest
  output <- mapply(function(x, y) {
    merge(select(x, sample, variable_to_be_classified),
      y,
      by = "sample",
      all.y = TRUE
    )
  },
  x = metadata_list,
  y = otu_cells_wide_list,
  SIMPLIFY = FALSE
  )


  return(output)
}

rf_variable <- "MeJA_treatment"
list_boruta_input <- physeq_to_borutaInput(ps_list_rarefied, rf_variable)


# check your df
str(list_boruta_input$Arabidopsis_thaliana.Root[1:10, 1:10]) # if your "rf_variable" is not a factor, boruta won't work

# run borura
RNGversion("4.0.2")
set.seed(456987)
boruta_objt_list <- lapply(list_boruta_input, function(x) {
  Boruta(MeJA_treatment ~ ., data = x, doTrace = 2, maxRuns = 500, ntree = 5000)
}) # increase the maximum number of runs to decrease the number of tenttively important OTUs. increase the number of trees to increase precision. decrease either to reduce computational time.
```

## 7.2 - visualize boruta results over a list (list as input, list as output)

```{r boruta}

# let Boruta decide if tentative features are ultimatetly important or not ;
fixed_boruta_objt_list <- lapply(boruta_objt_list, function(x) {
  TentativeRoughFix(x)
})

# get a list of ASVs defined as inportant by Boruta
boruta_ASV_list <- lapply(fixed_boruta_objt_list, function(x) {
  getSelectedAttributes(x)
})

# get the list of ASVs defined as inportant by Boruta in formula format ; this can be used to calculate precision
boruta_formula_list <- lapply(fixed_boruta_objt_list, function(x) {
  getConfirmedFormula(x)
})

# make a plot showing imporance of features
boruta_plot <- lapply(fixed_boruta_objt_list, function(x) {
  plot(x)
})

# make a plot showing imporance and classification of features over iteration time
boruta_history <- lapply(fixed_boruta_objt_list, function(x) {
  plotImpHistory(x)
})
```





##7.3 - Cheking Boruta precision with 100-repeated 5-fold cross-validation, over the list of 4 borutized random forested
```{r boruta}

############################ 3
# 100-repeated 5-fold cross-validation
#############################




#################### borutized forest
set.seed(4551)
boruta_feature_rf_repeatedcv <- mapply(function(x, z) {
  train.control <- trainControl(
    method = "repeatedcv", # set trainig/data split controls for the train function
    number = 5, repeats = 100
  )

  model_borutized <- train(z,
    data = x, method = "rf", # execute training based on random forest; model is based on borut formula of important features
    trControl = train.control, ntree = 1000
  )



  return(model_borutized)
},
x = list_boruta_input,
z = boruta_formula_list,
SIMPLIFY = FALSE
)

# check confusion matrix
map(boruta_feature_rf_repeatedcv, confusionMatrix)
```





## 7.4 - add the rf classification as part of the OTU taxonomy
```{r boruta}
##############################################################################################################
# add the rf classification as part of the OTU taxonomy
##############################################################################################################


# to a list of dataframes
rf_added_tax <- mapply(function(x, y) {
  tax_table(x) %>% # get tax table of the phyloseq object...
    as.data.frame() %>% # turn it into a dataframe so mutate can work
    mutate(
      random_forest_relevance = # make anew variable caled random_forest_relevance, where...
        if_else(taxa_names(x) %in% y == TRUE, "important", "uninportant")
    )
}, # %>% # if the taxa names are present in the boruta list of selected ASVs, tag them as inportant
# tax_table(), # make it a phyloseq object again #this still needs adjustment
x = ps_list_rarefied,
y = boruta_ASV_list,
SIMPLIFY = FALSE
)

# export to add to network visualizations
write.csv(select(rf_added_tax$Arabidopsis_thaliana.Soil, random_forest_relevance), "at_soil_rfimp.csv", row.names = TRUE)
write.csv(select(rf_added_tax$Brassica_oleraceae.Soil, random_forest_relevance), "bo_soil_rfimp.csv", row.names = TRUE)
write.csv(select(rf_added_tax$Arabidopsis_thaliana.Root, random_forest_relevance), "at_root_rfimp.csv", row.names = TRUE)
write.csv(select(rf_added_tax$Brassica_oleraceae.Root, random_forest_relevance), "bo_root_rfimp.csv", row.names = TRUE)
```



#7.6 - barplot of rf features
```{r}


# get boruta stats of ASVs confirmed to be important
rf_importance_byOTU <- lapply(fixed_boruta_objt_list, function(x) {
  filter(attStats(x), decision == "Confirmed") %>%
    tibble::rownames_to_column(var = "OTU")
})

#make ps object that only contains RF-imporantt ASVs
rf_important_ps_l<-mapply(function (z,y)
  prune_taxa(taxa = z$OTU,x = y),
  z=rf_importance_byOTU,
  y=ps_list_rarefied,
  SIMPLIFY = FALSE)

# turn ps object into a dataframe
melted_ps <- lapply(rf_important_ps_l, psmelt)

# join melted ps object created above with the boruta stats
rf_ASVs_df <- mapply(function(x, y) {
  left_join(x, y)
},
x = melted_ps,
y = rf_importance_byOTU,
SIMPLIFY = FALSE
)

# setup a new color pallete
color_treatments <- hcl.colors(4, palette = "Geyser")
color_treatments <- c("#E6C186", "#B8CDAE", "#008585", "#C7522B")

# quick plot bar with relative abudances, for a quick overview of  data
plot_rf_asvs <- lapply(rf_ASVs_df, function(x) {
  ggplot(data = x, aes(x = OTU, y = Abundance, fill = MeJA_treatment)) +
    geom_boxplot(aes()) +
    scale_fill_manual(values = color_treatments) +
    theme_bw() +
    theme(axis.title = element_text(size = 13, face = "bold")) +
    theme(
      panel.border = element_blank(), panel.grid.major = element_blank(),
      panel.grid.minor = element_blank(), axis.line = element_line(colour = "black")
    ) +
    theme(axis.text = element_text(size = 10, face = "bold", colour = "black")) +
    theme(axis.title.x = element_blank()) +
    geom_line(data = x, aes(x = OTU, y = meanImp), size = 2) +
    theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1))
})
```

# 7.7 - export Figures
```{r}
library(gridExtra)
plot_rf <- grid.arrange(plot_rf_asvs$Arabidopsis_thaliana.Root, 
                        plot_rf_asvs$Arabidopsis_thaliana.Soil,
                        plot_rf_asvs$Brassica_oleraceae.Root,
                        plot_rf_asvs$Brassica_oleraceae.Soil, 
                        nrow = 2, 
                        ncol = 2)


# save plot
ggsave(plot_rf,
  height = 9, width = 18,
  file = "./R output/Figures_for_publication/RF_ASVs_AT_BO.svg"
)
```



# 7.8 - overwrite saved session
```{r}

save.image("./R output/analysis_session.RData")

```


Random forest analysis complete! proceed to script 8_joint_ASV_classifications
