==> Code/9_important_taxa_overrepresentation.Rmd <==
---
title: "9_important_taxa_overrepresentation"
author: "Pedro"
date: "2/7/2022"
output: html_document
editor_options: 
  chunk_output_type: console
---

# load libraries and data
```{r}
source("./Code/Functions/load_libraries.R")
load("./R output/env.1_Loading_and_pre_processing.RData")
```



# 9.0 important taxa overrepresentation
Here we take the output of script 8 that defines teh ASVs that were importnat, and then look into the taxonomies of important taxa. It tests if  these taxa are overperesented in the importnat subset, when compared to the full subset. 

This script will awnser the following question: if 6 out of 17 taxa in RF_importnat of B.Ole roots are from f__Commonadacea, and 211 out of 2453 of the taxas in B.Ole roots are from f__Commonadacea, are these proportions similar or different?

# 9.1 important taxa overrepresentation
set a list of important taxa based on the list of important ASVs
```{r}


#list of ASVs tagged as important (chunk # 8.6.1b )
simplified_ASV_importance_scores

# generate a table to count occurences of tags across samples
imp_asv_table<-simplified_ASV_importance_scores
imp_asv_table<-lapply(imp_asv_table, function(x) x[,c(-2, -5)]) # remove info on neutral models and sum of tags
imp_asv_table<-lapply(imp_asv_table, function(x) x[rowSums(x)>0,]) # remove remove unimportat ASVs
imp_asv_table_colsum<-lapply(imp_asv_table, function(x) colSums(x)) # get sum of columns




#make table
imp_asv_table_colsum<-rbind(imp_asv_table_colsum[[1]],
  imp_asv_table_colsum[[2]],
  imp_asv_table_colsum[[3]],
  imp_asv_table_colsum[[4]])%>%as.data.frame()

#adjust rownmaes fo table
rownames(imp_asv_table_colsum)<-names(simplified_ASV_importance_scores)

#count number of ASVs with more than 1 tag, then add it to previous table
two_tags<-lapply(imp_asv_table, function(x) count(rowSums(x)>1))%>%unlist()%>%as.vector()
imp_asv_table_colsum$ASVs_with_two_tags<-two_tags


#total ASVs
n_taxa_imp<-lapply(imp_asv_table, function(x) dim(x)[1])%>%unlist()%>%as.vector()
imp_asv_table_colsum$n_taxa_imp<-n_taxa_imp


# get the ASV names of any ASV that had any tag (that is, remove uninportnat ASVS)
important_ASV_list<-lapply(simplified_ASV_importance_scores, function(x)
  row.names(x[x$sum_of_tags>0,]))

#as taxas above expected in neutral definitions were not defined in root samples but were explored in other figures, let's remove them from this analysis
important_ASV_list<-lapply(simplified_ASV_importance_scores, function(x){
  x<-x[,c(-2, -5)]
  x$sum_of_tags<-rowSums(x)
  row.names(x[x$sum_of_tags>0,])
})

# this is the total number of unique ASVs that had any tag  
important_ASV_list%>%unlist()%>%unique%>%length()

# total number of ASVs that had a network tag (derreplicated names
lapply(simplified_ASV_importance_scores, function(x)
  row.names(x[x$has_network_tag >0,]))%>%unlist()%>%unique()%>%length()


# this is the total number of unique ASVs that had any tag  
important_ASV_list%>%unlist()%>%unique%>%length()


# make list by hand as mapply fails us
ps_imp_taxa_l<-list(
  subset_taxa(ps_list_rarefied$Arabidopsis_thaliana.Root, 
              (taxa_names(ps_list_rarefied$Arabidopsis_thaliana.Root) %in% important_ASV_list$Arabidopsis_thaliana.Root)),
  
  subset_taxa(ps_list_rarefied$Arabidopsis_thaliana.Soil, 
              (taxa_names(ps_list_rarefied$Arabidopsis_thaliana.Soil) %in% important_ASV_list$Arabidopsis_thaliana.Soil)),
  
  subset_taxa(ps_list_rarefied$Brassica_oleraceae.Root, 
              (taxa_names(ps_list_rarefied$Brassica_oleraceae.Root) %in% important_ASV_list$Brassica_oleraceae.Root)),
  
  subset_taxa(ps_list_rarefied$Brassica_oleraceae.Soil, 
                        (taxa_names(ps_list_rarefied$Brassica_oleraceae.Soil) %in% important_ASV_list$Brassica_oleraceae.Soil)))
# assign names to list
names(ps_imp_taxa_l)<-c("Arabidopsis_thaliana.Root",
                         "Arabidopsis_thaliana.Soil",
                         "Brassica_oleraceae.Root",
                        "Brassica_oleraceae.Soil")


```




# 9.2 prepare data and calculate proportions with fisher's exact test

note: There was a problem on the loop when it is called inside the function: objects "i" in the loop were literally passed as charather "i", so classes/orders could not be found. the same code works fine when hard-coded. this was a problem when using phyloseq::subset_taxa(), but not now when using phyloseq::prune_taxa() 

```{r}
# source the function that calculates fisher tests for every taxonomic level
source("./Code/Functions/fisher_test_on_taxa_levels.R") 

# run the custom function voer 2 lists of philoseq objects, one with imporntat taxa and other with the full taxa (for every partition)
fisher_result_l<-mapply(function (x,y)
                   fisher_all_taxa_groups(ps_important_taxa = x, ps_all_taxa = y),
                   x = ps_imp_taxa_l,
                   y = ps_list_rarefied,
                   SIMPLIFY = FALSE)

#quick plot showing proportions of alphaproteobacteria and mucilaginibacter in BO_soil
#alphaprotteoac
alphaprot_DF<-c(
ntaxa(subset_taxa(ps_imp_taxa_l$Brassica_oleraceae.Soil, Class =="c__Alphaproteobacteria")),
ntaxa(ps_imp_taxa_l$Brassica_oleraceae.Soil)-ntaxa(subset_taxa(ps_imp_taxa_l$Brassica_oleraceae.Soil, Class =="c__Alphaproteobacteria")),
ntaxa(subset_taxa(ps_list_rarefied$Brassica_oleraceae.Soil, Class =="c__Alphaproteobacteria")),
ntaxa(ps_list_rarefied$Brassica_oleraceae.Soil)-ntaxa(subset_taxa(ps_list_rarefied$Brassica_oleraceae.Soil, Class =="c__Alphaproteobacteria")))
alphaprot_DF<-as.data.frame(alphaprot_DF)
alphaprot_DF$subset<-c("important", "important", "non-important", "non-important")
alphaprot_DF$taxa<-c("Alphaproteobacteria", "other", "Alphaproteobacteria", "other")

ggplot(alphaprot_DF, aes(y = alphaprot_DF, x = subset, fill = taxa))+
  geom_col(position = "fill")+
  theme_bw()+
  xlab(label = "ASV importance group")+
  ylab(label = "Proportion of Alphaproteobacteria")+
  theme(legend.position = "top")+
scale_fill_manual(values=c("#F8766D", "#00BFC4"))+
  geom_text(aes(label = alphaprot_DF), vjust = 1.1, position = "fill", color = "white")

  



#mucilaginibac
mucilag_DF<-c(
ntaxa(subset_taxa(ps_imp_taxa_l$Brassica_oleraceae.Soil, Genus =="g__Mucilaginibacter")),
ntaxa(ps_imp_taxa_l$Brassica_oleraceae.Soil),
ntaxa(subset_taxa(ps_list_rarefied$Brassica_oleraceae.Soil, Genus =="g__Mucilaginibacter")),
ntaxa(ps_list_rarefied$Brassica_oleraceae.Soil))
mucilag_DF<-as.data.frame(mucilag_DF)
mucilag_DF$subset<-c("important", "important", "non-important", "non-important")
mucilag_DF$taxa<-c("Mucilaginibacter", "other", "Mucilaginibacter", "other")

ggplot(mucilag_DF, aes(y = mucilag_DF, x = subset, fill = taxa))+
  geom_col(position = "fill")+
  theme_bw()+
  xlab(label = "ASV importance group")+
  ylab(label = "Proportion of Mucilaginibacter")+
  theme(legend.position = "top")+
scale_fill_manual(values=c("#F8766D", "#00BFC4"))+
  geom_text(aes(label = mucilag_DF), vjust = 1.1, position = "fill", color = "white")





#mucilaginibac
Comamonadaceae_DF<-c(
ntaxa(subset_taxa(ps_imp_taxa_l$Brassica_oleraceae.Soil, Family =="f__Comamonadaceae")),
ntaxa(ps_imp_taxa_l$Brassica_oleraceae.Soil)-ntaxa(subset_taxa(ps_imp_taxa_l$Brassica_oleraceae.Soil, Family =="f__Comamonadaceae")),
ntaxa(subset_taxa(ps_list_rarefied$Brassica_oleraceae.Soil, Family =="f__Comamonadaceae")),
ntaxa(ps_list_rarefied$Brassica_oleraceae.Soil)-ntaxa(subset_taxa(ps_list_rarefied$Brassica_oleraceae.Soil, Family =="f__Comamonadaceae")))
Comamonadaceae_DF<-as.data.frame(Comamonadaceae_DF)
Comamonadaceae_DF$subset<-c("important", "important", "non-important", "non-important")
Comamonadaceae_DF$taxa<-c("Comamonadaceae", "other", "Comamonadaceae", "other")

ggplot(Comamonadaceae_DF, aes(y = Comamonadaceae_DF, x = subset, fill = taxa))+
  geom_col(position = "fill")+
  theme_bw()+
  xlab(label = "ASV importance group")+
  ylab(label = "Proportion of Alphaproteobacteria")+
  theme(legend.position = "top")+
scale_fill_manual(values=c("#F8766D", "#00BFC4"))+
  geom_text(aes(label = Comamonadaceae_DF), vjust = 1.1, position = "fill", color = "white")




  


```


# 9.3 extract p values and odds ratio
Now that we have large lists of fisher tests, let's fish out important metrics for visualization
```{r}

#output
fisher_result_l



# non-adjusted p values for each fisher comparison
lapply(fisher_result_l, function(x) map(x,1))

# confidence interval of odds ratio for each fisher comparison
lapply(fisher_result_l, function(x) map(x,2))

# odds ratio for each fisher comparison
lapply(fisher_result_l, function(x) map(x,3))



#apply FDR correction to the p values
fdr_p_fisher_l<-lapply(fisher_result_l, function (x){
  #single df with all p values
  pvector<-do.call(rbind.data.frame, map(x,1))
  
  # adjust the p value with fdr, then save as a list
  adjusted_p<-p.adjust(p = pvector[[1]],
                          method = "fdr")%>%
                      as.list()
  # adjust names
  names(adjusted_p)<-names(x)

  return(adjusted_p)
  
})



# subset of FDR adjusted significant p values 
lapply (fdr_p_fisher_l, function (x) 
  x[x<0.05])

min3_adjustp<-str(lapply (fdr_p_fisher_l, function (x) 
  x[x<0.1]))



```



# 9.4 make a heat tree of important taxa
metacoder heat trees are made to visualize differences in taxonomic composition, so let's incoporate p and odds ratio values into such trees
```{r}

fisher_to_heatTree<-function(fisher_output_l,
                             ps_important_taxa_l){
#######################################
######## make metacoder object ######## 
#######################################
imp_heat_ps<-ps_important_taxa_l

#remove unecessary taxonomic indo (dada2id, "S__" and" above_selected)
tax_table(imp_heat_ps)<-tax_table(imp_heat_ps)[,1:6]

# let's remove the "r__"ranks from the taxonomy, they can be useful but will polute our plot
tax_table(imp_heat_ps)[, colnames(tax_table(imp_heat_ps))] <- gsub(tax_table(imp_heat_ps)[, colnames(tax_table(imp_heat_ps))],     pattern = "[a-z]__", replacement = "")


# transform from phyloseq to  taxmap object
imp_heat<-parse_phyloseq(imp_heat_ps)



#######################################
######## aplly fdr to fisher ######## 
#######################################

#apply FDR correction to the p values

  #single df with all p values
  pvector<-do.call(rbind.data.frame, map(fisher_output_l,1))
  
  # adjust the p value with fdr, then save as a list
  adjusted_p<-p.adjust(p = pvector[[1]],
                          method = "fdr")%>%
                      as.list()
  # adjust names
  names(adjusted_p)<-names(fisher_output_l)

  

#this is the name that mustch match our fisher p list
taxon_id_metacoder<-lapply(imp_heat$taxa, function (x)
                                                    x$get_name())%>%
                                                    map(1)


# now turn that list into a df
taxon_id_metacoder<-do.call(rbind.data.frame, map(taxon_id_metacoder,1))

# and change column name
colnames(taxon_id_metacoder)<-"taxa_id"

# truns the p values into a dataframe
adjusted_p_input<-as.data.frame(t(as.data.frame(adjusted_p)))%>%
  rownames_to_column()

#change column names
colnames(adjusted_p_input)<-c("taxa_id", "fdr_p")


# let's remove the "r__"ranks from the taxonomy, they can be useful but will polute our plot
adjusted_p_input$taxa_id<- gsub(adjusted_p_input$taxa_id,     
                                pattern = "[a-z]__", 
                                replacement = "")

  
# now join both df
fdr_for_heatTree<-left_join(taxon_id_metacoder,
                            adjusted_p_input,
                            by = "taxa_id")



#######################################################################
################### add fold changes! ########################################
#######################################################################


# get odds ratio for each taxa
odd_list<- map(fisher_output_l,3)%>%
  map(1)

# truns the p values into a dataframe
odd_list<-as.data.frame(t(as.data.frame(odd_list)))%>%
  rownames_to_column()

#change column names
colnames(odd_list)<-c("taxa_id", "odds_ratio")

# let's remove the "r__"ranks from the taxonomy, they can be useful but will polute our plot
odd_list$taxa_id<- gsub(odd_list$taxa_id,     
                        pattern = "[a-z]__", 
                        replacement = "")


# now join both df
fisher_for_heatTree<-left_join(fdr_for_heatTree,
                            odd_list,
                            by = "taxa_id")




#NAs will crash metadore. let's change them with 0.9999 for now
fisher_for_heatTree[is.na(fisher_for_heatTree)]<-"0.99999"

#change structure to numeric
fisher_for_heatTree$fdr_p<-as.numeric(fisher_for_heatTree$fdr_p)
fisher_for_heatTree$odds_ratio<-as.numeric(fisher_for_heatTree$odds_ratio)


#now let's dim p values above 0.1, or else we can get counfounded
#fisher_for_heatTree$fdr_p[fisher_for_heatTree$fdr_p>0.1]<-1







#######################################################################
############### now plot the heat tree ############### 
#######################################################################



set.seed(1)
      output<- heat_tree(imp_heat,
                 node_size = fisher_for_heatTree$odds_ratio, # n_obs is a function that calculates, in this case, the number of OTUs per taxon
                 node_color = fisher_for_heatTree$fdr_p,
                 node_label = taxon_names,
                 node_size_axis_label = "Size: oods_ratio",
                 node_color_interval = c(0, 0.15),
                 edge_color_interval = c(0, 0.15),
                 node_color_range = c("darkgreen", "seagreen", "yellowgreen", "grey"),
                 edge_color_range = c("darkgreen", "seagreen", "yellowgreen", "grey"),
                 node_color_axis_label = "Color: FDR-adjusted p",
                 layout = "davidson-harel", # The primary layout algorithm
                 initial_layout = "reingold-tilford") # The layout algorithm that initializes node locations

      return(output)
}


# add the fisher values to the heat trees
output_fisher_summary_trees<-mapply(function(x,y)
    fisher_to_heatTree(x,y),
    x = fisher_result_l,
    y = ps_imp_taxa_l,
    SIMPLIFY = FALSE)

#make a single 4-pannel plot with A-D labels
output_fisher_summary_trees<-ggarrange(output_fisher_summary_trees$fisher_result_AT_root,
                                        output_fisher_summary_trees$fisher_result_AT_soil,
                                        output_fisher_summary_trees$fisher_result_BO_root,
                                        output_fisher_summary_trees$fisher_result_BO_soil,
                                        common.legend = TRUE,
                                        labels = c("A)", "B)", "C)", "D)"))

#save plot
ggsave(output_fisher_summary_trees,
       height = 22, width = 22,
       file="./R output/Figures_for_publication/fisher_fourpartiitons.svg")


```



# overwrite saved session
```{r}

save.image("./R output/analysis_session.RData")

```


analysis complete!
